refactor: update LiveScoreHandler and related classes to use ScheduleHandler and ScoringHandler for improved clarity and maintainability
