+++ src/main/java/org/thingai/app/scoringservice/dto/ScoreDetailDto.java
@@ -1,14 +1,14 @@
 package org.thingai.app.scoringservice.dto;
 
 import org.thingai.app.scoringservice.entity.score.Score;
-import org.thingai.app.scoringservice.handler.ScoreHandler;
+import org.thingai.app.scoringservice.handler.ScoringHandler;
 
 public class ScoreDetailDto {
     private Score baseScore;
     private Object detailScore;
 
     public ScoreDetailDto() {
-        baseScore = ScoreHandler.factoryScore();
+        baseScore = ScoringHandler.factoryScore();
     }
 
     public void setScore(String allianceId) {
