+++ src/main/java/org/thingai/app/scoringservice/handler/MatchControlHandler.java
@@ -1,0 +1,47 @@
+package org.thingai.app.scoringservice.handler;
+
+public class MatchControlHandler {
+    private static final String TAG = "MatchControlHandler";
+    private static final int MATCH_DURATION_SECONDS = 180;
+
+    public MatchControlHandler() {
+
+    }
+
+    // Match control methods
+    public void loadMatch(String matchId) {
+
+    }
+
+    public void activateMatch(String matchId) {
+    }
+
+    public void startMatch() {
+
+    }
+
+    public void abortMatch() {
+
+    }
+
+    public void commitScore() {
+
+    }
+
+    public void overrideScore() {
+
+    }
+
+    // Display control methods
+    public void showPreview() {
+
+    }
+
+    public void showMatch() {
+
+    }
+
+    public void postScoreResults(String matchId) {
+
+    }
+}
